Updated the data_final filtering set so that it uses the datetime and went through the 2024 data manually to check which ones set off the error for get_TIA_grd.
I added in notes to clarify the importance of the GAP number in GPS_track() which stops analysis happening between gaps wider than 8hours.

Added notes from message from Chris T about the reason why there is an error for some birds when get_TIA_grid is run.

Added in an if catch for birds that have enough tracks for the TIA but most tracks are outside the UK so don't have a habitat layer (e.g. 9L spent most of the time in France but had one point in spring transition in the UK).