Added in code to as.POSIXct the time columns for 1 day, 1 week, 2 weeks and 6 weeks. Without this some of the < assessments were coming up false when they should have been true so not all fixes were in the correct time periods.
So in order to make this work in the loop I have hash tagged out the convert to BTOTT track object until in the loop so that the date columns stay as.POSIXct.

So added the correct rows into each of the different compartments of the loop. Will carry on tomorrow