A few additional tweaks
the past_cohort_behavs spreadsheet read in needed a :%S to include 59s of the time.

Realised one more bird needed filtering from the dataset = 9J from winter post breeding 2024 - so added that into the excludeall file.

Removed the trial ifelse loops from last year now that the code is working properly.

Additional code to check how many fixes were beyond the 7th Dec 2024 which is when the bouy sensor stopped working.

Added the remaining turn it into a Tracl 'BTOTT' convert code at the end of each of the individual time periods and updated that loop so that it works on an if else {} else if {} properly. Added the female loop into the last else.

Moved the set seed to above the final loop