Additional bird time period to exclude - "0E" for end of december 2022 (winter post breeding).
Because of this removal there is no data for 2022 winter post breeding. So the code associated with this time period for 2022 has been removed at the various points.

Small tweak to the habitat plots so that the key is above the bars for the cohort graphs.

Added in code to full join and then go wide on the AUC and Beta data tables so that it makes it easier to copy over from excel for the report.