A few more tweaks to the code.
Added in a loop for the tide figure - for 2024 for July-Dec 2024 to filter the data fixes that are after the 7th Dec 2024 because the sensor broke. So the figure needs to be created from the data without those fixes.

Added in two more lines of code to extract the overall F, DFs and p values from the lm() model runs. And added this to the table of mytest.contrast to be saved out as a csv.